Move HTMLUtils class to util package
HTMLUtils methods are used in MosesTranslate.
This make dependency from MT connector to Filter driver.
This indicate the class is common utility so
refactor the class into org.omegat.util package.

- Rename it to org.omegat.util.HtmlUtils
- Deprecate org.omegat.filters2.html2.HTMLUtils class
  and refer to HtmlUtils for compatibility.
