+++ src/org/omegat/filters2/hhc/HHCFilterVisitor.java
@@ -30,7 +30,7 @@ import java.io.BufferedWriter;
 
 import org.htmlparser.Tag;
 import org.omegat.filters2.html2.FilterVisitor;
-import org.omegat.filters2.html2.HTMLUtils;
+import org.omegat.util.HtmlUtils;
 
 /**
  * The part of HTML filter that actually does the job. This class is called back
@@ -80,8 +80,8 @@ class HHCFilterVisitor extends FilterVisitor {
     protected void maybeTranslateAttribute(Tag tag, String key) {
         String attr = tag.getAttribute(key);
         if (attr != null) {
-            String trans = filter.privateProcessEntry(HTMLUtils.entitiesToChars(attr), null);
-            tag.setAttribute(key, HTMLUtils.charsToEntities(trans, filter.getTargetEncoding(), sShortcuts));
+            String trans = filter.privateProcessEntry(HtmlUtils.entitiesToChars(attr), null);
+            tag.setAttribute(key, HtmlUtils.charsToEntities(trans, filter.getTargetEncoding(), sShortcuts));
         }
     }
 }
